Move shared logger to train logger
