+++ tests/conftest.py
@@ -21,8 +21,8 @@ Command = list[str]
 
 
 from zeroband.training.data import STABLE_FILE
+from zeroband.training.logger import reset_logger
 from zeroband.training.world_info import reset_world_info
-from zeroband.utils.logger import reset_logger
 from zeroband.utils.models import AttnImpl
 from zeroband.utils.parquet import pa_schema
